requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=3.14.0
openai>=0.27.0
python-dotenv>=1.0.0
